feat: Chat settings dropdown
